git-svn-id: http://springhead.info/spr2/Springhead2/trunk@6204 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Creature/CRBone.h
@@ -203,7 +203,7 @@ public:
 
 	/** @brief 通過点を数を取得する
 	*/
-	int NTrajectoryNodes() { return trajNodes.size(); }
+	int NTrajectoryNodes() { return (int)(trajNodes.size()); }
 
 	/** @brief 時刻tにおけるの通過点を取得する
 	*/
